refactor(protocol): move discriminator module into opportunity/ domain
packages/protocol/src/pool/ -> packages/protocol/src/opportunity/discriminator/

The pool being mined is the candidate pool of *opportunities* (discovery-run
output rows), so the module belongs in the opportunity domain next to its
relatives (question.generator.ts, discovery-question.helper.ts, feed/).
Mirrors the existing opportunity/feed/ subdirectory pattern. Import paths
only; no behavior change; 30/30 tests green.
