+++ packages/protocol/src/opportunity/discriminator/discriminator.scorer.ts
